insurance_web/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
Pillow==10.2.0
python-magic==0.4.27
requests==2.31.0
gunicorn==21.2.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 